[rename] : 폴더 구조&이름 변경
