+++ client/src/hooks/useScrollToTop.ts
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import { useLocation } from 'react-router-dom';
 
-export default function ScrollToTop() {
+export default function useScrollToTop() {
 	const { pathname } = useLocation();
 
 	useEffect(() => {
