+++ client/src/pages/Community/PostUpdate.tsx
@@ -1,7 +1,7 @@
-import '@toast-ui/editor/dist/toastui-editor.css';
-import '../../Global.css';
-import 'tui-color-picker/dist/tui-color-picker.css';
 import '@toast-ui/editor-plugin-color-syntax/dist/toastui-editor-plugin-color-syntax.css';
+import '@toast-ui/editor/dist/toastui-editor.css';
+import 'tui-color-picker/dist/tui-color-picker.css';
+import '../../Global.css';
 
 import { ChangeEvent, KeyboardEvent, useEffect, useRef, useState } from 'react';
 
@@ -13,10 +13,10 @@ import colorSyntax from '@toast-ui/editor-plugin-color-syntax';
 /* eslint-disable jsx-a11y/label-has-associated-control */
 import { Editor } from '@toast-ui/react-editor';
 
-import { Api } from '../../apis/customAPI';
-import useAxios from '../../apis/customAxios';
 import SearchPlace from '../../Components/Community/SearchPlace';
 import SubjectDropdown from '../../Components/Community/SubjectDropdown';
+import { Api } from '../../apis/customAPI';
+import useAxios from '../../hooks/useAxios';
 
 const Container = styled.div`
 	width: 100vw;
