+++ client/src/pages/Community/ReiviewDetail.tsx
@@ -11,11 +11,11 @@ import Swal from 'sweetalert2';
 
 import { Viewer } from '@toast-ui/react-editor';
 
-import { Api } from '../../apis/customAPI';
-import useAxios from '../../apis/customAxios';
 import Answers from '../../Components/Community/Answers';
 import MapApi from '../../Components/Community/MapApi';
 import ReviewCarousel from '../../Components/Community/ReviewCarousel';
+import { Api } from '../../apis/customAPI';
+import useAxios from '../../hooks/useAxios';
 import { Iuser } from '../../reducers/userInfoReducer';
 import { RootState } from '../../store/Store';
