+++ client/src/pages/common/Landing.tsx
@@ -1,9 +1,9 @@
-import '../Global.css';
+import '../../Global.css';
 
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 
-import logo from '../assets/logo.png';
+import logo from '../../assets/logo.png';
 
 const ScrollSnapWrap = styled.div`
 	scroll-snap-type: y mandatory;
