+++ client/src/pages/common/MainPage.tsx
@@ -5,14 +5,14 @@ import { useSelector } from 'react-redux';
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 
-import Banner from '../Components/Banner';
-import Carousel from '../Components/Carousel';
-import CarouselNearbyPlace from '../Components/CarouselNearbyPlace';
-import CarouselReview from '../Components/CarouselReview';
-import MainHeader from '../Components/MainHeader';
-import UserHeader from '../Components/UserHeader';
-import { Ilogin } from '../reducers/loginReducer';
-import { RootState } from '../store/Store';
+import Banner from '../../Components/mainpage/Banner';
+import Carousel from '../../Components/mainpage/Carousel';
+import CarouselNearbyPlace from '../../Components/mainpage/CarouselNearbyPlace';
+import CarouselReview from '../../Components/mainpage/CarouselReview';
+import MainHeader from '../../Components/mainpage/MainHeader';
+import UserHeader from '../../Components/mainpage/UserHeader';
+import { Ilogin } from '../../reducers/loginReducer';
+import { RootState } from '../../store/Store';
 
 const MainContainer = styled.div`
 	width: 100vw;
