+++ client/src/pages/contents/HotReview.tsx
@@ -3,7 +3,7 @@ import { HTMLAttributes, useEffect, useState } from 'react';
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
 
-import useAxios from '../apis/customAxios';
+import useAxios from '../../hooks/useAxios';
 
 interface SlideItemProps extends HTMLAttributes<HTMLDivElement> {
 	image?: string;
