+++ client/src/pages/contents/NearbyPlace.tsx
