+++ client/src/pages/contents/RegionDetail.tsx
