+++ client/src/pages/member/Mypage.tsx
@@ -7,10 +7,10 @@ import { Link, useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
 import Swal from 'sweetalert2';
 
-import { Api } from '../../apis/customAPI';
-import useAxios from '../../apis/customAxios';
 import IntroBox from '../../Components/member/IntroBox';
 import MyReview from '../../Components/member/MyReview';
+import { Api } from '../../apis/customAPI';
+import useAxios from '../../hooks/useAxios';
 import { Iuser, UPDATE } from '../../reducers/userInfoReducer';
 import { RootState } from '../../store/Store';
